tests/wpt/tests/navigation-api/navigation-methods/reload-base-url.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="resources/page-with-base-url-common.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    const startingURL = new URL(i.src);
    i.contentWindow.navigation.reload();
    i.onload = t.step_func_done(() => {
      const iframeURL = new URL(i.contentWindow.navigation.currentEntry.url);
      assert_equals(iframeURL.pathname, startingURL.pathname);
    });
  });
}, "reload() must ignore the Navigation object's base URL");
</script>
